improving python scripts

+++ simple/compile.py
@@ -6,7 +6,17 @@ def appendToFile(filename, content):
 	f.write(content + "\n")
 	f.close()
 
-def main():
+
+def is_integer(value):
+	try:
+		int(value)
+		return True
+	except ValueError:
+		print "Unable to parse the input!"
+		return False
+
+
+def generate_main_body(main_bench, google_bench):
 	ahf="python ~/acchelper/acchelper.py forward"
 	ahb="python ~/acchelper/acchelper.py backward"
 
@@ -14,38 +24,38 @@ def main():
 	google_bench_lib_path  	= "-L/home/millad/benchmark/bin/lib"
 	google_bench_lib 		= "-lbenchmark"
 
-	os.system("rm -rf bin")
-	os.system("rm -rf bin-googlebench")
+	if main_bench:
+		os.system("rm -rf bin")
+		os.system("mkdir -p bin")
+		os.system("cd src;" + ahf + ";cd ..;")
 
-	os.system("mkdir -p bin")
-	os.system("mkdir -p bin-googlebench")
-
-	os.system("cd src;" + ahf + ";cd ..;")
-	os.system("cd src-googlebench;" + ahf + ";cd ..;")
+	if google_bench:
+		os.system("rm -rf bin-googlebench")
+		os.system("mkdir -p bin-googlebench")
+		os.system("cd src-googlebench;" + ahf + ";cd ..;")
 
 	os.system("rm -f log-compile.log")
 	os.system("rm -f compile-output.log")
 
-	pattern = None
-	if len(sys.argv) >= 2:
-		pattern = sys.argv[1]
+	pattern_list = []
 
-	print "pattern: %s" % (pattern)
+	if main_bench:
+		pattern_list.append("src/*.cpp")
 
-	for pattern in ["src/*.cpp", "src-googlebench/*.cpp"]:
-	# for pattern in ["src-googlebench/*.cpp"]:
+	if google_bench:
+		pattern_list.append("src-googlebench/*.cpp")
+
+
+	for pattern in pattern_list:
 
 		filelist = []
 
-		# if pattern is None:
-		# 	filelist = glob.glob("./src/*.cpp")
-		# else:
 		filelist = glob.glob(pattern)
 
 		filelist = sorted(filelist)
 
 
-		compile_cmd = "pgc++ -std=c++11 -w -O3 -mp -acc -ta=tesla%s %s %s %s -o %s %s &>> log-compile.log"
+		compile_cmd = "pgc++ -std=c++11 -w -O3 -mp -acc -ta=tesla%s %s %s %s -o %s %s >> log-compile.log 2>&1"
 
 		for f in filelist:
 			managed = ":cc70,cc60"
@@ -66,11 +76,42 @@ def main():
 			os.system(cmd)
 			appendToFile("compile-output.log", cmd)
 
-	os.system("cd src;" + ahb + ";cd ..;")
-	os.system("cd src-googlebench;" + ahb + ";cd ..;")
+	if main_bench:
+		os.system("cd src;" + ahb + ";cd ..;")
+
+	if google_bench:
+		os.system("cd src-googlebench;" + ahb + ";cd ..;")
+
 
 	print "Error count:"
 	os.system("wc -l log-compile.log")
+
+
+def main():
+
+	while True:
+		print "\n\nWhich source files to generate:"
+		print "1- Main source files"
+		print "2- Google Benchmark files"
+		print "3- Both"
+		print "0- Exit"
+
+		inp = raw_input("Your choice: ")
+		print "---------"
+		if is_integer(inp):
+			inp = int(inp)
+			if inp == 0:
+				break
+			elif inp == 1:
+				generate_main_body(True, False)
+			elif inp == 2:
+				generate_main_body(False, True)
+			elif inp == 3:
+				generate_main_body(True, True)
+			else:
+				print "Please choose from above options."
+				continue
+			break
